feat(ci): Add PyPI publishing job to GitLab CI
Introduces a new  job to the GitLab CI  stage. This job automates the process of building and publishing the package to PyPI upon a successful tagged release (vX.Y.Z) on the  branch.

The job utilizes Poetry for building and publishing, and requires  and  CI/CD variables to be configured.